+++ test/associated-model.js
@@ -130,8 +130,6 @@ $(document).ready(function () {
         return found ? found : id;
     };
 
-    Employee = {};
-
     Employee = Backbone.AssociatedModel.extend({
         relations:[
             {
@@ -148,7 +146,7 @@ $(document).ready(function () {
             {
                 type:Backbone.One,
                 key:'manager',
-                relatedModel:Employee
+                relatedModel:'Employee'
             }
         ],
         validate:function (attr) {
@@ -877,6 +875,72 @@ $(document).ready(function () {
         f.at(0).get('xxx').add({name:"n2"}); //Fire add:xxx
     });
 
+    test("Polymorphic Associate : Issue#23", 4, function () {
+
+        var Models = {};
+        var findPolyMorphicType = Models.findPolyMorphicType = function (relation, attributes) {
+            var key = relation.key + '_type';
+            return Models[attributes[key] || this.get(key)];
+        };
+
+        Models.Job = Backbone.AssociatedModel.extend({
+            relations:[
+                {
+                    type:Backbone.One,
+                    key:'organizable',
+                    relatedModel:findPolyMorphicType
+                }
+            ]
+
+        });
+
+        Models.Comment = Backbone.AssociatedModel.extend({
+            relations:[
+                {
+                    type:Backbone.One,
+                    key:'commentable',
+                    relatedModel:findPolyMorphicType
+                }
+            ]
+        });
+
+        Models.Company = Backbone.AssociatedModel.extend({
+
+        });
+        Models.Department = Backbone.AssociatedModel.extend({
+
+        });
+
+
+        Models.Article = Backbone.AssociatedModel.extend({
+
+        });
+        Models.StatusUpdate = Backbone.AssociatedModel.extend({
+
+        });
+
+        var cjob = new Models.Job({organizable_type:'Company', name:"J1", organizable:{name:"Google"}});
+        var djob = new Models.Job({organizable_type:'Department', name:"J2", organizable:{name:"Google Reader"}});
+
+        var articleComment = new Models.Comment({
+            commentable_type:'Article',
+            name:"c1",
+            commentable:{body:"Wonderful post!"}
+        });
+        var statusComment = new Models.Comment({
+            commentable_type:'StatusUpdate',
+            name:"c2",
+            commentable:{body:"Why are you updating your status with pointless crap?"}
+        });
+
+        equal(cjob.get('organizable') instanceof Models.Company, true);
+        equal(djob.get('organizable') instanceof Models.Department, true);
+
+        equal(articleComment.get('commentable') instanceof Models.Article, true);
+        equal(statusComment.get('commentable') instanceof Models.StatusUpdate, true);
+
+    });
+
 
     test("Fetch collection with reset: Issue#45", 5, function () {
         var Product = Backbone.AssociatedModel.extend({
@@ -1414,6 +1478,44 @@ $(document).ready(function () {
         emp2.get('manager').set({'fname':'newEmp2'});
         equal(emp2.get('fname'), 'newEmp2', "emp's fname should be changed");
         equal(emp2.get('manager').get('fname'), 'newEmp2', "manager's fname should be changed");
+
+        var emp3 = new Employee({fname:'emp3', manager:{fname:'emp4'}});
+        equal(emp3.get('manager.fname'), 'emp4', "manager's fname should be emp4");
+    });
+
+    test("Backbone.Self", 9, function () {
+        var User = Backbone.AssociatedModel.extend({
+            relations:[
+                {
+                    key:'friends',
+                    type:Backbone.Many,
+                    relatedModel:Backbone.Self
+                }
+            ],
+            defaults:{
+                username:undefined,
+                fname:'',
+                lname:'',
+                aboutMe:'',
+                friends:[]
+            }
+        });
+
+        var user1 = new User({id:1, username:'user1'});
+        var user2 = new User({id:2, username:'user2', friends:[user1]});
+        var user3 = new User({id:3, username:'user3', friends:[user1, {id:4, username:'user4'}]});
+
+        equal(user1.get('username'), 'user1', "user1's username should be correct");
+        equal(user1.get('friends').length, 0, "count of friends of user1 should be 0");
+
+        equal(user2.get('username'), 'user2', "user2's username should be correct");
+        equal(user2.get('friends').length, 1, "count of friends of user2 should be 1");
+        equal(user2.get('friends[0].username'), 'user1', "username of first follower of user2 should be user1");
+
+        equal(user3.get('username'), 'user3');
+        equal(user3.get('friends').length, 2);
+        equal(user3.get('friends[0].username'), 'user1');
+        equal(user3.get('friends[1].username'), 'user4');
     });
 
     test("Self-Reference `toJSON`", function () {
@@ -1532,7 +1634,7 @@ $(document).ready(function () {
         });
     });
 
-    test("parents", 6, function () {
+    test("parent relations", 7, function () {
         emp.set('works_for', {name:"Marketing", number:29});
 
         var emp2 = new Employee({
@@ -1569,13 +1671,13 @@ $(document).ready(function () {
         equal(emp.parents.length, 0);
         equal(emp2.parents.length, 0);
 
-        console.log(works_for.parents);
-
-        //Could cause mem leaks
+        //Could cause mem leaks if you set emp3 = undefined
         //emp3 = undefined;
         //equal(works_for.parents.length,1);
-        //console.log(works_for.parents);
 
+        emp3.cleanup();
+        emp3 = undefined;
+        equal(works_for.parents.length, 0);
 
     });