+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+### Version 0.5.1 - [Diff](https://github.com/dhruvaray/backbone-associations/compare/v0.5.0...v0.5.1)
+
+* `relatedModel` can now also be a function. Could be used to model polymorphic associations like in this [gist](https://gist.github.com/dhruvaray/5988996).
+* `relatedModel` can be passed a special value called `Backbone.self` to handle self references.
+* Introducing explicit cleanup on `Backbone.Model` to prevent memory leaks when users set `Backbone.Model` to undefined explicitly.
+* Fixed the handling of {reset:true} option on `Backbone.Collection`.
+* Fixed a scenario where repeated fetch on the same `Backbone.Model` or `Backbone.Collection` caused event bubbling to stop.
+
+
 #### Version 0.5.0 - [Diff](https://github.com/dhruvaray/backbone-associations/compare/v0.4.2...v0.5.0)
 * Introducing "nested-change" event to allow parents (and higher) to listen to any changes in the hierarchy.
 * Introducing * operator in paths involving collections. Has semantics of any item in the collection.